trying to make irqs work (no success)
interestingly, the IRQs themselves actually *do* get triggered!
but for some reason the vector (just like most stuff on the stack)
is equal to 0xfafafafa, which I can't really explain why.
why 0xfafafafa? is this some sick kind of joke? why does it happen as
soon as I `asm volatile("sti");`? and this number is everywhere,
including the result of `inl(0x60);` function. i need to look into
that...