GDT issues?
after enabling interrupts the CPU starts sending #GP (0x0d) exceptions
and setting some registers to 0xfafafafa. the latter only happened to
me, but looking through QEMU logs gave me an idea that there might be an
issue with GDT.
apparently, the error code I get when the CPU gets all screwed up is the
key to finding the cause of this problem. 0x200 is 0b1000000000 in
binary, and after some research it turns out that that there's an issue
with... entry 0x40 in the GDT??? now, that's a bit too suspicious,
though something always happens when handling interrupts, and GDT[0x40]
might be one of them.
please send help