interrupts are working properly! mostly, at least
it turns out that I am an idiot (what a surprise!)
the reason for messed up vectors and errors is that I forgot to put the
dollar sign before integer literals, and they have, in turn, been
treated as addresses, thus giving weird values.
next up is to fix weirdly occurring PIC IRQs that are triggered by while
loops.