FINALLY! WORKING IRQs!
now the keyboard is a lot smoother and faster, the timer is working
(though it's not used), and the IRQs do not bring the entire system down.
also, for the time being, I think it's best that exceptions halt the CPU
as it is painful to recover from them (at least for me).